Grafana

+++ postgres/init.sql
@@ -4,7 +4,7 @@ CREATE DATABASE metrics;
 CREATE TABLE IF NOT EXISTS rate(
     ID                serial          PRIMARY KEY,
     created_at        DATE            NOT NULL,
-    coast             smallint        NOT NULL,
+    coast             int        NOT NULL,
     stonks            bool            default false,
     coins             varchar(255)    NOT NULL
 )
